#322 [fix] 검색 전용 meeting 서비스 생성 및 회의 시간 함수 로직 이동
meeting service 내에 많은 관심사가 있어서
조회용 서비스 파일을 만들어서 최적의 회의 시간을 추천하는 함수를 옮겼습니다
